+++ tests/unit_tests/test_etree_element_helpers.py
