+++ tests/unit_tests/test_mig.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Tuple
 import pytest  # type:ignore[import]
 
 from maus.models.message_implementation_guide import SegmentGroupHierarchy, SegmentGroupHierarchySchema
-from unit_tests.serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
+from serialization_test_helper import assert_serialization_roundtrip  # type:ignore[import]
 
 ALL_SGH_FILES = pytest.mark.datafiles(
     "./migs/FV2204/segment_group_hierarchies/sgh_mscons.json",
